+++ Libplanet.Tests/HashSetExtensions.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 namespace Libplanet.Tests
 {
-    public static class HashSetExtension
+    public static class HashSetExtensions
     {
         public static HashSet<T> ToHashSet<T>(this IEnumerable<T> l)
         {